fix: resolve ultracode branch-review findings across engine, 3D, and editor (003)
Address 10 confirmed findings from the 6-dimension adversarial branch review:

- D1-F04: cache pixel-budget uses long arithmetic to prevent int overflow at high density
- D1-F02: capture op.Bounds before Dispose to eliminate use-after-dispose
- D1-F01: sanitize MaxWorkingScale in RenderNodeProcessor (mirrors OutputScale guard)
- D1-F03: add VerifyAccess to DrawRenderTargetScaled / DrawImageScaled
- D2-F01: MosaicEffect Relative origin uses output buffer dimensions (DeviceBufferSize)
- D3-F01: DrawableTextureSource forwards density as maxWorkingScale to inner processor
- D3-F02: Renderer3D.Resize uses build-then-commit pattern (mirrors Initialize)
- D5-F02: IRenderer3D.SurfaceDensity is now get-only; setter stays on Renderer3D
- D5-F12: EffectTarget.Scale changed to { get; init; }
- D4-F01: SaveFrameDialogViewModel implements IDisposable with CompositeDisposable

+++ src/Beutl.Engine/Graphics/Rendering/Cache/RenderNodeCacheHelper.cs
@@ -125,11 +125,11 @@ public static class RenderNodeCacheHelper
         }
 
         var list = processor.RasterizeToRenderTargets(ops);
-        int pixels = list.Sum(i =>
+        long pixels = list.Sum(i =>
         {
             // Budget the ACTUAL stored pixels (density-scaled), matching the allocated tile size.
             var pr = outputScale == 1f ? PixelRect.FromRect(i.Bounds) : PixelRect.FromRect(i.Bounds, outputScale);
-            return pr.Width * pr.Height;
+            return (long)pr.Width * pr.Height;
         });
         if (!cacheOptions.Rules.Match(pixels))
         {
@@ -188,11 +188,11 @@ public readonly record struct RenderCacheRules(int MaxPixels, int MinPixels)
 
     public bool Match(PixelSize size)
     {
-        int count = size.Width * size.Height;
+        long count = (long)size.Width * size.Height;
         return MinPixels <= count && count <= MaxPixels;
     }
 
-    public bool Match(int pixels)
+    public bool Match(long pixels)
     {
         return MinPixels <= pixels && pixels <= MaxPixels;
     }
